+++ lib/fluent/fluentui.dart
@@ -66,6 +66,7 @@ Widget buildFluentUI(BuildContext context) {
   SystemChrome.setSystemUIOverlayStyle(const SystemUiOverlayStyle(
     systemNavigationBarColor: Colors.transparent,
     systemNavigationBarDividerColor: Colors.transparent,
+    systemNavigationBarContrastEnforced: false,
     statusBarColor: Colors.transparent,
   ));
   final botToastBuilder = BotToastInit();
@@ -126,9 +127,8 @@ Widget buildFluentUI(BuildContext context) {
             value: SystemUiOverlayStyle(
               systemNavigationBarColor: Colors.transparent,
               systemNavigationBarDividerColor: Colors.transparent,
+              systemNavigationBarContrastEnforced: false,
               statusBarColor: Colors.transparent,
-              systemNavigationBarContrastEnforced: true,
-              systemStatusBarContrastEnforced: true,
               systemNavigationBarIconBrightness: brightness == Brightness.dark
                   ? Brightness.light
                   : Brightness.dark,