+++ plugins/rhttp/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.18.0
+
+- feat: restore previous behavior of trusting root certificates from webpki-roots broken by 0.17.0
+- deps: bump Rust dependencies to latest versions
+- **BREAKING**: `trustRootCertificates` changed to `rootCertSource`
+
+## 0.17.0
+
+- deps: bump `reqwest` to `0.13.3` (Might be breaking! See README for Android setup, no longer needed in 0.18.0) #105
+
 ## 0.16.0
 
 - deps: bump `flutter_rust_bridge` to `2.12.0`